requirements_cu118.txt
--extra-index-url https://download.pytorch.org/whl/cu118

numpy==1.26.4
opencv-python==4.10.0.84
scikit-image==0.21.0
pillow==9.5.0
onnx==1.16.1
protobuf==4.25.3
psutil==6.0.0
onnxruntime-gpu==1.18.0
packaging==24.1
PySide6==6.8.2.1
kornia
torch==2.4.1+cu118
torchvision==0.19.1+cu118
torchaudio==2.4.1+cu118
tqdm
ftfy
regex
pyvirtualcam==0.11.1
tensorrt-cu11==10.4.0
numexpr
onnxsim
requests
PyYAML>=6.0.1
pyqt-toast-notification==1.3.2
qdarkstyle
pyqtdarktheme